Fixes get_tier_for_org
Basically django doesn't support cross-db relations. Which is
understandable, but what this gives us is the inability to query for a
Tier objects since the construction of the related Organization will
fail, since it's trying to construct a edx-organization, rather than a
amc organization which is what is saved in the db that we're querying.

We fix this by using .values() which doesn't try to construct an
organization model. We then use the dictionary to make a Tier object
with the organization field pointing to self (edx-organization)

Bumps version